# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Team Titans - Free Lunch API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/auth/user/signin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User/Organization login",
                "operationId": "b25c088b399a97e86d1156957fb2aae3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@example.com",
                                    "password": "1Password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "User/Organization login response",
                                        "value": {
                                            "message": "User authenticated successfully",
                                            "statusCode": 200,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE_ENTITY",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "User/Organization login response",
                                        "value": {
                                            "message": [
                                                "field is required",
                                                "field is required"
                                            ],
                                            "statusCode": 422,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "User/Organization login response",
                                        "value": {
                                            "message": "Authentication failed",
                                            "statusCode": 403,
                                            "status": "error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/user/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset user password",
                "operationId": "ed6cd9e3e9b14386ed6324d44f52e990",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@example.com",
                                    "note": "Thank you for the good work"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Reset user password",
                                        "value": {
                                            "message": "OTP sent successfully",
                                            "statusCode": 200,
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE_ENTITY",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Reset user password",
                                        "value": {
                                            "message": [
                                                "[Field] is required",
                                                "[Field] is required"
                                            ],
                                            "statusCode": 422
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Reset user password",
                                        "value": {
                                            "message": "User not found",
                                            "statusCode": 404
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}